requirements.txt
# Requirements for CrimeTime.
# To use type: pip install -r requirements.txt

## Requirements without Version specifiers

shapely==1.6.4
flask==1.0.2
python-dateutil==2.7.5
pandas==0.23.4
numpy==1.15.4
geopandas==0.3.0
geopy==1.17.0
folium==0.7.0
matplotlib==3.0.1
tornado==5.1.1
jupyter==1.0.0
descartes==1.1.0
beautifulsoup4==4.6.3
pytest==4.0.0
sphinx==1.8.2
